update WebUtilService interface
move some tests for WebUtil from ApiServiceSpec
