+++ rundeckapp/src/test/groovy/rundeck/services/ApiServiceSpec.groovy
@@ -32,6 +32,7 @@ import groovy.util.slurpersupport.GPathResult
 import groovy.xml.MarkupBuilder
 import org.grails.plugins.codecs.JSONCodec
 import org.rundeck.app.authorization.AppAuthContextEvaluator
+import org.rundeck.app.web.WebUtilService
 import org.rundeck.core.auth.AuthConstants
 import org.springframework.context.MessageSource
 import rundeck.AuthToken
@@ -633,74 +634,12 @@ class ApiServiceSpec extends HibernateSpec implements ControllerUnitTest<ApiCont
 
     }
 
-    @Unroll
-    def "render error json"() {
-        given:
-        service.messageSource = Mock(MessageSource) {
-            getMessage(_, _,_, _) >> { it[0] + (it[1] ?: '') }
-        }
-        when:
-        def result = service.renderErrorJson(data, code)
-
-        then:
-        println result
-        result == '{"error":true,"apiversion":' +
-                (ApiVersions.API_CURRENT_VERSION) +
-                        (expectCode ? ',"errorCode":"' + (expectCode) +'"' : '') +
-                        (eMessage ? ',"message":"' + (eMessage) +'"' : '') +
-                        (eMessages ? ',"messages":' + (eMessages) : '') +
-                        '}'
-
-
-        where:
-        code   | data                                 | expectCode          | eMessage            | eMessages
-        null   | [:]                                  | 'api.error.unknown' | 'api.error.unknown' | null
-        'test' | [:]                                  | 'test'              | 'api.error.unknown' | null
-        'test' | ['a', 'b']                           | 'test'              | null                | '["a","b"]'
-        'test' | [code: 'acode']                      | 'test'              | 'acode'             | null
-        'test' | [code: 'acode', message: 'amessage'] | 'test'              | 'amessage'          | null
-        'test' | [code: 'acode', args: ['a', 'b']]    | 'test'              | 'acode[a, b]'       | null
-        'test' | [message: 'amessage']                | 'test'              | 'amessage'          | null
-
-    }
-
-    @Unroll
-    def "render error xml"() {
-        given:
-        service.messageSource = Mock(MessageSource) {
-            getMessage(_, _,_, _) >> { it[0] + (it[1] ?: '') }
-        }
-        when:
-        def result = service.renderErrorXml(data, code)
-        then:
-        result == """<result error='true' apiversion='${ApiVersions.API_CURRENT_VERSION}'>
-  <error code='${
-            code
-        }'>
-    """+
-                (emessage?("<message>${emessage}</message>"):'')+
-                (elist?("""<messages>\n      ${elist}\n    </messages>"""):'')+
-        """
-  </error>
-</result>"""
-
-        where:
-        code    | data                                 | emessage            | elist
-        'acode' | null                                 | 'api.error.unknown' | null
-        'acode' | [code: 'acode']                      | 'acode'             | null
-        'acode' | [code: 'acode', args: ['a', 'b']]    | 'acode[a, b]'       | null
-        'acode' | [code: 'acode', message: 'zmessage'] | 'zmessage'          | null
-        'acode' | [message: 'amessage']                | 'amessage'          | null
-        'acode' | ['a', 'b']                           | null                | '''<message>a</message>\n      <message>b</message>'''
-
-
-    }
-
     def "render success xml response unwrapped"() {
-
+        given:
+        service.rundeckWebUtil= Mock(WebUtilService)
         when:
         def closureCalled = false
-        def result = service.renderSuccessXml(response) {
+        service.renderSuccessXml(response) {
             closureCalled = true
             responseData(test:true){
                 value('something')
@@ -708,60 +647,15 @@ class ApiServiceSpec extends HibernateSpec implements ControllerUnitTest<ApiCont
         }
 
         then:
+            1 * service.rundeckWebUtil.respondOutput(_, 'application/xml', _) >> {
+                def slurper = new XmlSlurper()
+                def gpath = slurper.parseText(it[2])
+                assert gpath.name() == 'responseData'
+                assert gpath.'@test'.text() == 'true'
+                assert gpath.value.text() == 'something'
+            }
         closureCalled
-        response.contentType == "application/xml"
-        response.characterEncoding == "UTF-8"
-        response.getHeader("X-Rundeck-API-Version") == ApiVersions.API_CURRENT_VERSION.toString()
-        response.getHeader("X-Rundeck-API-XML-Response-Wrapper") == null
-        def slurper = new XmlSlurper()
-        def gpath = slurper.parseText(response.text)
-        gpath.name()=='responseData'
-        gpath.'@test'.text()=='true'
-        gpath.value.text()=='something'
-    }
-
-    def "requre version invalid xml"() {
-        given:
-        service.messageSource = Mock(MessageSource) {
-            getMessage(_, _,_, _) >> { it[0] +":"+ (it[1].join(":") ?: '') }
-        }
-        when:
-        request.addHeader('accept','application/xml')
-        def check = service.requireVersion([api_version:1,forwardURI: '/test/uri'],response,2)
-
-        then:
-        def result = assertXmlErrorText(response.text)
-        !check
-        response.contentType == "application/xml"
-        response.characterEncoding == "UTF-8"
-        response.status == 400
-        response.getHeader("X-Rundeck-API-Version") == ApiVersions.API_CURRENT_VERSION.toString()
-        result.error.message.text() == 'api.error.api-version.unsupported:1:/test/uri:Minimum supported version: 2'
-    }
-    def "requre version invalid default json"() {
-        given:
-        service.messageSource = Mock(MessageSource) {
-            getMessage(_, _,_, _) >> { it[0] +":"+ (it[1].join(":") ?: '') }
-        }
-        when:
-        def check = service.requireVersion([api_version:1,forwardURI: '/test/uri'],response,2)
-
-        then:
-        !check
-        response.contentType == "application/json"
-        response.characterEncoding == "UTF-8"
-        response.status == 400
-        response.getHeader("X-Rundeck-API-Version") == ApiVersions.API_CURRENT_VERSION.toString()
-        response.json.message == 'api.error.api-version.unsupported:1:/test/uri:Minimum supported version: 2'
     }
 
 
-    private GPathResult assertXmlErrorText(String result) {
-        def slurper = new XmlSlurper()
-        def gpath = slurper.parseText(result)
-        'result' == gpath.name()
-        'true' == gpath['@error'].text()
-        ApiVersions.API_CURRENT_VERSION.toString() == gpath['@apiversion'].text()
-        gpath
-    }
 }
